+++ src/flx_repl.hpp
@@ -20,7 +20,7 @@ namespace interpreter {
 	class FlexaRepl {
 	public:
 		static void remove_header(std::string& err);
-		static void count_scopes(const std::string& input_line, unsigned int& open_scopes);
+		static void count_scopes(const std::string& input_line, size_t& open_scopes);
 		static std::string read(const std::string& msg);
 		static int execute(const FlexaCliArgs& args);
